requirements.txt
flask
google-genai
pymongo
